raft::read_barrier, unavailable leader handling
If the leader is unavailable during read_barrier, closed_error occurs, which was not handled in any way and eventually reached the client. This patch adds retries in this case.

The patch also adds logging of server errors occuring while processing a request, as otherwise it is difficult to debug problems.

Fix: scylladb/scylla-dtest#2753