raft_topology: fence RPCs when transitioning to tokens_state::read_write
The patch introduces the concept of topology version,
an int which is incremented when the topology state
changes. It's used to catch and block with error the RPCs
which were issued based on the old topology version
when the cluster is transitioning to the new one.

The patch adds topology version handling to the storage proxy.
The main function try_apply_fence compares the version
from the RPC parameters with the current one. If the version
from the parameters is greater than the current version, we
wait for the topology update on the current node
(wait_for_next_version). If the version from the parameters is
less than the current version, and the range no longer belongs
to the current node, an exception is thrown. At the moment,
this exception causes the original request to fail, in the future
the code will be improved so that the original request is
retried on the new topology.