raft_initialize_discovery_leader: use raft_timeout{}
This function is called as part of a node startup
procedure, so a timeout may be useful.

As outlined in the comment, there is no valid way
we can lose quorum here, but some subsystems may
just become unreasonably slow for various reasons,
so we nonetheless use raft_timeout{} here.