query_options: add replica_local mode
We want to access the paxos state table only on the local node and
shard (or shards in case of intranode_migration). In this commit we
add a replica_local flag to query_options, which allows to do that.
This flag can be set for a query via make_internal_options.

We handle this flag on the statements layer, in select_statement
and modification_statement. This